Мелкие исправления.

+++ trunk/PointsBot/Random.h
@@ -27,9 +27,9 @@ inline ulong Random()
 #if RAND_MAX == 0x7FFF // В этом случае rand() генерирует 15 случайных бит. Смещениями можно получить 32/64 бита.
 
 #if ULONG_MAX == 0xFFFFFFFFUL
-	return ((ulong)rand() << 17) | ((ulong)rand() << 2) | ((ulong)rand() & 3);
+	return (rand() << 17) | (rand() << 2) | (rand() & 3);
 #elif ULONG_MAX == 0xFFFFFFFFFFFFFFFFUL
-	return ((ulong)rand() << 49) | ((ulong)rand() << 34) | ((ulong)rand() << 19) | ((ulong)rand() << 4) | ((ulong)rand() & 15);
+	return (rand() << 49) | (rand() << 34) | (rand() << 19) | (rand() << 4) | (rand() & 15);
 #else
 #error Unknown ULONG_MAX
 #endif
@@ -37,9 +37,9 @@ inline ulong Random()
 #elif RAND_MAX == 0x7FFFFFFF // В этом случае rand() генерирует 31 случайный бит. Смещениями можно получить 32/64 бита.
 
 #if ULONG_MAX == 0xFFFFFFFFUL
-	return ((ulong)rand() << 1) | ((ulong)rand() & 1);
+	return (rand() << 1) | (rand() & 1);
 #elif ULONG_MAX == 0xFFFFFFFFFFFFFFFFUL
-	return ((ulong)rand() << 33) | ((ulong)rand() << 2) | ((ulong)rand() & 3);
+	return (rand() << 33) | (rand() << 2) | (rand() & 3);
 #else
 #error Unknown ULONG_MAX
 #endif
